move receipt tests to their own file
